switch the Atom's d_mass member back to a double; the float was causing too many problems with downstream systems.

+++ Code/GraphMol/Atom.h
@@ -527,7 +527,7 @@ namespace RDKit{
     atomindex_t d_index;
     boost::uint16_t d_isotope;
 
-    float d_mass;
+    double d_mass;
     ROMol *dp_mol;
     Dict *dp_props;
     AtomMonomerInfo *dp_monomerInfo;
